+++ MasterFramework Dev/Framework/Drawing/DrawingGroup.lua
@@ -68,10 +68,8 @@ end
          - `drawingGroup:CachedSize()`: Returns the width, height computed in the last `drawingGroup:UpdateLayout()`
          - `drawingGroup:AbsolutePosition()`: Returns the x, y coordinates the drawing group will use to draw on-screen
 ]]
-function framework:DrawingGroup(body, disableDrawList)
+function framework:DrawingGroup(body)
     local drawingGroup = {}
-
-    drawingGroup.disableDrawList = disableDrawList or Internal.debugMode.disableDrawList
 
     drawingGroup.drawers = {}
     drawingGroup.drawTargets = {}
@@ -262,7 +260,7 @@ function framework:DrawingGroup(body, disableDrawList)
             gl_Translate(absoluteX, absoluteY, 0)
         end
 
-        if self.disableDrawList or next(self.continuouslyUpdatingDrawers) then
+        if Internal.debugMode.disableDrawList or next(self.continuouslyUpdatingDrawers) then
             _Draw(self)
 
             for drawer, _ in pairs(self.continuouslyUpdatingDrawers) do
@@ -307,9 +305,10 @@ function framework:DrawingGroup(body, disableDrawList)
     function drawingGroup:DrawerUpdated(drawer)
         if element 
         and isWithinViewport 
-        and self.drawers[drawer]
-        and not (self.disableDrawList or next(self.continuouslyUpdatingDrawers)) then
-            element.requestedRedraws[redrawFunc] = true
+        and self.drawers[drawer] then
+            if not (Internal.debugMode.disableDrawList or next(self.continuouslyUpdatingDrawers)) then
+                element.requestedRedraws[redrawFunc] = true
+            end
             return true
         end
     end
@@ -332,7 +331,7 @@ function framework:DrawingGroup(body, disableDrawList)
     function drawingGroup:DrawerWillNotContinuouslyUpdate(drawer)
         if self.drawers[drawer] then
             self.continuouslyUpdatingDrawers[drawer] = nil
-            if not (self.disableDrawList or next(self.continuouslyUpdatingDrawers)) then
+            if not (Internal.debugMode.disableDrawList or next(self.continuouslyUpdatingDrawers)) then
                 element.requestedRedraws[redrawFunc] = true
             end
             return true