+++ changelog.md
@@ -55,6 +55,7 @@ Bug fixes:
 - Hide `MovableFrame` handle when releasing outside bounds)
 - Fix over-cropping for `OffsettedViewport`. (I suspect this happens with height too, but so far I haven't been able to verify. There's chance the error's something else, idk.)
 - Fix bug where, when a font stopped being used in a `TextGroup`, it would cancel drawing for all following fonts.
+- Fix continuous updating being incompatible with calling `drawingGroup:DrawerUpdated(drawer)`
 
 Also includes further non-breaking (hopefully) optimisations and fixes.